+++ assembly/src/ast/parsers/io_ops.rs
@@ -9,6 +9,13 @@ use crate::{StarkField, ADVICE_READ_LIMIT, HEX_CHUNK_SIZE, MAX_PUSH_INPUTS};
 use core::{convert::TryFrom, ops::RangeBounds};
 use vm_core::WORD_SIZE;
 
+/// Helper enum for endianness determination in the parsing functions.
+#[derive(Debug)]
+enum Endianness {
+    Little,
+    Big,
+}
+
 // CONSTANTS
 // ================================================================================================
 
@@ -33,7 +40,7 @@ pub fn parse_push(op: &Token, constants: &LocalConstMap) -> Result<Node, Parsing
             match param_str.strip_prefix("0x") {
                 // if we have only one hex parameter
                 Some(param_str) if param_str.len() <= HEX_CHUNK_SIZE => {
-                    let value = parse_hex_value(op, param_str, 1)?;
+                    let value = parse_hex_value(op, param_str, 1, Endianness::Big)?;
                     build_push_one_instruction(value)
                 }
                 // if we have many hex parameters without delimiter
@@ -249,7 +256,7 @@ fn parse_param_list(op: &Token, constants: &LocalConstMap) -> Result<Node, Parsi
     let values =
         op.parts().iter().enumerate().skip(1).map(|(param_idx, &param_str)| {
             match param_str.strip_prefix("0x") {
-                Some(param_str) => parse_hex_value(op, param_str, param_idx),
+                Some(param_str) => parse_hex_value(op, param_str, param_idx, Endianness::Big),
                 None => parse_non_hex_param_with_constants_lookup(
                     op,
                     constants,
@@ -281,32 +288,29 @@ fn parse_non_hex_param_with_constants_lookup<R: RangeBounds<u64>>(
     }
 }
 
-/// Parses a single hexadecimal parameter into multiple values and returns an appropriate push
+/// Parses a 64-character hex string into a word (4 field elements) and returns an appropriate push
 /// instruction node.
 ///
 /// # Errors
 /// Returns an error if:
-/// - The length of hex string is not even.
-/// - The length of hex string is not divisible by 16.
+/// - The length of hex string is not equal to 64.
 /// - If the string does not contain a valid hexadecimal value.
 /// - If the parsed value is greater than or equal to the field modulus.
-fn parse_long_hex_param(op: &Token, param_str: &str) -> Result<Node, ParsingError> {
+fn parse_long_hex_param(op: &Token, hex_str: &str) -> Result<Node, ParsingError> {
     // handle error cases where the hex string is poorly formed
-    if param_str.len() % HEX_CHUNK_SIZE != 0 {
+    if hex_str.len() != HEX_CHUNK_SIZE * WORD_SIZE {
         // hex string doesn't contain a valid number of bytes
         return Err(ParsingError::invalid_param_with_reason(
             op,
             1,
-            &format!(
-                "hex string '{param_str}' does not contain a number of characters multiple 16"
-            ),
+            &format!("long hex string '{hex_str}' must contain exactly 64 characters"),
         ));
     }
 
     // iterate over the multi-value hex string and parse each 8-byte chunk into a valid u64
-    let values = (0..param_str.len())
+    let values = (0..hex_str.len())
         .step_by(HEX_CHUNK_SIZE)
-        .map(|i| parse_hex_value(op, &param_str[i..i + HEX_CHUNK_SIZE], 1));
+        .map(|i| parse_hex_value(op, &hex_str[i..i + HEX_CHUNK_SIZE], 1, Endianness::Little));
 
     build_push_many_instruction(values)
 }
@@ -315,35 +319,57 @@ fn parse_long_hex_param(op: &Token, param_str: &str) -> Result<Node, ParsingErro
 ///
 /// # Errors
 /// Returns an error if:
-/// - The length of hex string is not even.
-/// - The length of hex string is greater than 16.
+/// - The length of a short hex string (big-endian) is not even.
+/// - The length of a short hex string (big-endian) is greater than 16.
+/// - The length of the chunk of a long hex string (little-endian) is not equal to 16.
 /// - If the string does not contain a valid hexadecimal value.
 /// - If the parsed value is greater than or equal to the field modulus.
-fn parse_hex_value(op: &Token, param_str: &str, param_idx: usize) -> Result<u64, ParsingError> {
-    if param_str.len() % 2 != 0 {
-        return Err(ParsingError::invalid_param_with_reason(
-            op,
-            param_idx,
-            &format!("hex string '{param_str}' does not contain an even number of characters"),
-        ));
-    }
-
-    if param_str.len() > HEX_CHUNK_SIZE {
-        return Err(ParsingError::invalid_param_with_reason(
-            op,
-            param_idx,
-            &format!("hex string '{param_str}' contains too many characters"),
-        ));
-    }
-
-    let value = u64::from_str_radix(param_str, 16)
-        .map_err(|_| ParsingError::invalid_param(op, param_idx))?;
+fn parse_hex_value(
+    op: &Token,
+    hex_str: &str,
+    param_idx: usize,
+    endianness: Endianness,
+) -> Result<u64, ParsingError> {
+    let value = match endianness {
+        Endianness::Big => {
+            if hex_str.len() % 2 != 0 {
+                return Err(ParsingError::invalid_param_with_reason(
+                    op,
+                    param_idx,
+                    &format!(
+                        "hex string '{hex_str}' does not contain an even number of characters"
+                    ),
+                ));
+            }
+            if hex_str.len() > HEX_CHUNK_SIZE {
+                return Err(ParsingError::invalid_param_with_reason(
+                    op,
+                    param_idx,
+                    &format!("hex string '{hex_str}' contains too many characters"),
+                ));
+            }
+            u64::from_str_radix(hex_str, 16)
+                .map_err(|_| ParsingError::invalid_param(op, param_idx))?
+        }
+        Endianness::Little => {
+            if hex_str.len() != HEX_CHUNK_SIZE {
+                return Err(ParsingError::invalid_param_with_reason(
+                    op,
+                    param_idx,
+                    &format!("hex string chunk '{hex_str}' must contain exactly 16 characters"),
+                ));
+            }
+            u64::from_str_radix(hex_str, 16)
+                .map(|v| v.swap_bytes())
+                .map_err(|_| ParsingError::invalid_param(op, param_idx))?
+        }
+    };
 
     if value >= Felt::MODULUS {
         Err(ParsingError::invalid_param_with_reason(
             op,
             param_idx,
-            &format!("hex string '{param_str}' contains value greater than field modulus"),
+            &format!("hex string '{hex_str}' contains value greater than field modulus"),
         ))
     } else {
         Ok(value)